Show Edit Interface Sections also for FB instances
The interface edit sections for typed subapps are now also shown for FB
instances. Harmonizing the look and feel and make space in the interface
section for pin member access (e.g., children of structs).
